Put behaviour tests and state tests into same class
